Use double instead of float

+++ src/Network.cpp
@@ -31,7 +31,7 @@ void Network::generate_weights()
 	}
 }
 
-float Network::sigmoid_activation(float x)
+double Network::sigmoid_activation(double x)
 {
     return 1.0/(1.0 + exp(-x));
 }
